Add more window windows managers
