+++ modules/platform/window_managers/windows/__init__.py
@@ -1,0 +1,3 @@
+from .factory import get_window_mgr_on_windows
+
+__all__ = ["get_window_mgr_on_windows"]
