+++ modules/platform/window_managers/windows/win32gui_manager.py
@@ -1,6 +1,6 @@
 import logging
 
-from .base import WindowMgr
+from ..base import WindowMgr
 
 log = logging.getLogger(__name__)
 
@@ -12,20 +12,12 @@ except ImportError:
     HAS_WIN32GUI = False
     log.warning("win32gui not installed")
 
-try:
-    from ahk import AHK
-
-    HAS_AHK = True
-except ImportError:
-    HAS_AHK = False
-    log.warning("AHK Not Installed")
-
 
 HEARHTSTONE_WINDOW_NAME_WINDOWS = "Hearthstone"
 SW_SHOW = 5
 
 
-class WindowMgrWindows(WindowMgr):
+class WindowMgrWindowsWin32Gui(WindowMgr):
     """Encapsulates some calls to the winapi for window management"""
 
     def __init__(self):
@@ -51,33 +43,3 @@ class WindowMgrWindows(WindowMgr):
     def _set_foreground(self):
         """put the window in the foreground"""
         win32gui.SetForegroundWindow(self._handle)
-
-
-class WindowMgrWindowsAHK(WindowMgr):
-    """Encapsulates some calls to AHK for window management"""
-
-    def __init__(self):
-        """Constructor"""
-        self._win = None
-        self._ahk = AHK()
-
-    def find_game(self):
-        """find the hearthstone game window"""
-        win = self._ahk.win_get(title=HEARHTSTONE_WINDOW_NAME_WINDOWS)
-        win.show()
-        win.to_top()
-        win.activate()
-        self._win = win
-        return self._win
-
-    def get_window_geometry(self):
-        return self._win.rect
-
-
-def get_window_mgr_on_windows():
-    if HAS_WIN32GUI:
-        return WindowMgrWindows
-    elif HAS_AHK:
-        return WindowMgrWindowsAHK
-    else:
-        log.error("No Window Manager found for Windows")
